Control.Monad.State.Class

Description

MonadState class.

This module is inspired by the paper /Functional Programming with Overloading and Higher-Order Polymorphism/, Mark P Jones (http://web.cecs.pdx.edu/~mpj/) Advanced School of Functional Programming, 1995.

Documentation

class Monad m => MonadState m where

get returns the state from the internals of the monad.

put replaces the state inside the monad.

Associated Types

type StateType m

Methods

get :: m (StateType m)

put :: StateType m -> m ()

modify :: MonadState m => (StateType m -> StateType m) -> m ()

Monadic state transformer.

Maps an old state to a new state inside a state monad. The old state is thrown away.

      Main> :t modify ((+1) :: Int -> Int)
      modify (...) :: (MonadState Int a) => a ()

This says that modify (+1) acts over any Monad that is a member of the MonadState class, with an Int state.

gets :: MonadState m => (StateType m -> a) -> m a

Gets specific component of the state, using a projection function supplied.