Control.Monad.Writer.Lazy

Description

Lazy writer monads.

Inspired by the paper /Functional Programming with Overloading and Higher-Order Polymorphism/, Mark P Jones (http://web.cecs.pdx.edu/~mpj/pubs/springschool.html) Advanced School of Functional Programming, 1995.

MonadWriter class

class (Monoid (WriterType m), Monad m) => MonadWriter m where

Associated Types

type WriterType m

Methods

tell :: WriterType m -> m ()

listen :: m a -> m (a, WriterType m)

pass :: m (a, WriterType m -> WriterType m) -> m a

listens :: MonadWriter m => (WriterType m -> b) -> m a -> m (a, b)

censor :: MonadWriter m => (WriterType m -> WriterType m) -> m a -> m a

The Writer monad

type Writer w = WriterT w Identity

A writer monad parameterized by the type w of output to accumulate.

The return function produces the output mempty, while >>= combines the outputs of the subcomputations using mappend.

runWriter :: Writer w a -> (a, w)

Unwrap a writer computation as a (result, output) pair. (The inverse of writer.)

execWriter :: Writer w a -> w

Extract the output from a writer computation.

execWriter m = snd (runWriter m)

mapWriter :: ((a, w) -> (b, w')) -> Writer w a -> Writer w' b

Map both the return value and output of a computation using the given function.

runWriter (mapWriter f m) = f (runWriter m)

The WriterT monad transformer

newtype WriterT w m a

A writer monad parameterized by:

w - the output to accumulate.

m - The inner monad.

The return function produces the output mempty, while >>= combines the outputs of the subcomputations using mappend.

Constructors

WriterT

Fields

runWriterT :: m (a, w)

execWriterT :: Monad m => WriterT w m a -> m w

Extract the output from a writer computation.

execWriterT m = liftM snd (runWriterT m)

mapWriterT :: (m (a, w) -> n (b, w')) -> WriterT w m a -> WriterT w' n b

Map both the return value and output of a computation using the given function.

runWriterT (mapWriterT f m) = f (runWriterT m)

module Control.Monad

module Control.Monad.Fix

module Control.Monad.Trans

module Data.Monoid